Refactored TestTimeReporter multiverse output
