+++ test/multiverse/lib/multiverse/runner.rb
@@ -47,11 +47,15 @@ module Multiverse
     end
 
     def run(filter = "", opts = {})
-      # This file is generated by TestTimeReporter. The class cannot be loaded
-      # until minitest has been loaded by an individual suite, so we cannot define
-      # this inside the TestTimeReporter class. Please keep this path and the
-      # path for TestTimeReporter::TEST_TIME_REPORT the same!
-      FileUtils.rm_f(Multiverse::TIME_REPORT_PATH)
+      # These files are generated by TestTimeReporter and SuiteTimeReporter.
+      # These classes cannot be loaded until minitest has been loaded by an
+      # individual suite, so we cannot define this inside the TestTimeReporter
+      # class. Please keep this path and the paths for the reporter constants
+      # the same!
+      # TODO: These files aren't connected to the correct path, so they aren't getting reset
+      File.delete(Multiverse::TEST_TIME_REPORT_PATH)
+      File.delete(Multiverse::SUITE_TIME_REPORT_PATH)
+
       execute_suites(filter, opts) do |suite|
         puts yellow(suite.execution_message)
         suite.each_instrumentation_method do |method|
